FIRE-10549 etc. part one of a potentially larger change to come later
This fixes the hop grid name issue. The expectation is that an HG grid has a valid GridURL parameter in the OpenSim extras, if not the home grid will be used as normal. Extensions to this such as map copy slurl, and other places may follow in a future release but require more investigation.