[BUG-232834][BUG-232871] Alpha mask is being applied with emissive mask
This is the most local fix for this issue, addressing the specific unqualified use of HAS_ALPHA_MASK.
If we find other issues with alpha mask being applied incorrectly then, it may be better to fix higher up in llvieweshadermgr.cpp by reverting the changes from SL-17532. 
For now, this way works for this specific bug without non-emissive side-effects.