FIRE-32905 - Allow a larger range of fps for limit
As 144 Hz is very common and 240 and even 360Hz are feasible it makes no sense to limit this at 120. I have made the upper limit 360 for now. At some point no doubt this will not be good enough but we may need to review the limiting code at that point as it will be increasingly inaccurate given the use of millisecond sleep