DRTVWR-587: Simplify Python.cmake by omitting find_package(Python3).
Elsewhere in CMake land, we reference PYTHONINTERP_FOUND and
PYTHON_EXECUTABLE, both of which are explicitly set by Python.cmake. We don't
seem to need the find_package(Python3 COMPONENTS Interpreter) call. Given that
we take some pains to be careful about which Windows Python interpreter we
find, this eliminates a wildcard.